telebot/requirements.txt
python-dotenv
aiogram
openai 